+++ src/parsers/index.ts
@@ -1,8 +1,0 @@
-/**
- * Service key parsers
- */
-
-export { IServiceKeyParser } from './IServiceKeyParser';
-export { AbapServiceKeyParser } from './AbapServiceKeyParser';
-export { XsuaaServiceKeyParser } from './XsuaaServiceKeyParser';
-